Added basic LLM functionality (#27)
Added SpeziLLM as a dependency to the project and created very basic
infrastructure for the chief complaint (a button on the home screen
leading to openAI onboarding and chat like what we saw in the demo on
Jan 25th).

# *Integrate basic LLM Functionality to the project*

## :recycle: Current situation & Problem
Our project heavily depends on LLMs, but we did not yet have SpeziLLM as
a dependency nor did we have any views set up for a chat interface with
the LLM.


## :gear: Release Notes 
*Added SpeziLLM (including SpeziLLMOpenAI and SpeziLLMLocal) as a
dependency to the project*
*Created a Create Form tab on the home screen that links to a basic chat
interface for the Chief Complaint*
*Initialized for OpenAI LLM functionality, but not yet ready for local
(i.e. no code written for downloading model)*


## :books: Documentation
The LLM modules are contained in the ChiefComplaint group for now.
LLMInteraction is the chat UI, LLMOnboarding is the first time setup
(for OpenAI, it requests an API Key, and once we get local LLM set up it
will be able to initiate the setup for that). Onboarding files for
OpenAI initialization are found in the OpenAI group, and onboarding
files for local initialization are found in the Local group.


## :white_check_mark: Testing
All the code in the ChiefComplaint group passes all the tests for a
successful build at compile time.


## :pencil: Code of Conduct & Contributing Guidelines 

By submitting creating this pull request, you agree to follow our [Code
of
Conduct](https://github.com/CS342/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/CS342/.github/blob/main/CONTRIBUTING.md):
- [Y] I agree to follow the [Code of
Conduct](https://github.com/CS342/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/CS342/.github/blob/main/CONTRIBUTING.md).