Llm chat/finalize (#41)
# *LLM Cheief Complaint Functionality, Prompt, and Summary View *

## :recycle: Current situation & Problem
Before the pull request, we had a minimal implementation of the LLM
chatbot. Our summary appeared but it looked ugly, and the LLM did not
ask very detailed or specific questions.


## :gear: Release Notes 
The LLM makes a summary of the chief complaint with a textbox that the
patient is able to edit. The summary is now much more robust, and the
LLM asks the patient more relevant questions than it did before. It's
not limited anymore to just duration and severity. The textbox is now
editable.


## :books: Documentation
We commented out the naviagtion stack for now and plan on incorporating
that after merging. We changed the LLM function call to now be more
non-parametric and dynamic. We got rid of the parameters and instructed
it to give us a summary instead. With less structure, we are able to
harness all of the LLM's abilities.

## :white_check_mark: Testing
We have tested the LLM on a variety of medical concerns that are not
included in the examples given in the system prompt, and we have gotten
really great results with specific questions to the complaint.


## :pencil: Code of Conduct & Contributing Guidelines 

By submitting creating this pull request, you agree to follow our [Code
of
Conduct](https://github.com/CS342/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/CS342/.github/blob/main/CONTRIBUTING.md):
- [x] I agree to follow the [Code of
Conduct](https://github.com/CS342/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/CS342/.github/blob/main/CONTRIBUTING.md).

---------

Co-authored-by: nriedman <108841122+nriedman@users.noreply.github.com>